Update parse.py
This is the preliminary parser, still kind of useless since i didn't test it yet. Some functionalities, such as strings, arrays, indexing has not been implemented. We can see whether we want them implemented or not. And of course, many things will have to change to accomodate the construction of the parse tree.